fix(frontend): resolve react-refresh lint error in AuthContext
Separate Context creation and useAuth hook into dedicated files to
satisfy react-refresh/only-export-components rule.
