+++ frontend/src/contexts/AuthContext.tsx
@@ -1,15 +1,6 @@
-import { createContext, useContext } from 'react'
-import type { User } from '@/types/api'
 import { useMe, useLogout } from '@/hooks/useAuth'
-
-interface AuthContextValue {
-  user: User | null | undefined
-  isLoading: boolean
-  isAuthenticated: boolean
-  logout: () => void
-}
-
-const AuthContext = createContext<AuthContextValue | null>(null)
+import { AuthContext } from '@/contexts/authContextValue'
+import type { AuthContextValue } from '@/contexts/authContextValue'
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
   const { data: user, isLoading } = useMe()
@@ -23,12 +14,4 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
   }
 
   return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>
-}
-
-export function useAuth() {
-  const context = useContext(AuthContext)
-  if (!context) {
-    throw new Error('useAuth must be used within AuthProvider')
-  }
-  return context
 }
