+++ frontend/src/hooks/useAuthContext.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react'
+import { AuthContext } from '@/contexts/authContextValue'
+
+export function useAuth() {
+  const context = useContext(AuthContext)
+  if (!context) {
+    throw new Error('useAuth must be used within AuthProvider')
+  }
+  return context
+}
